refactor(hash table): hide implementation information that is unnecessary to the user
